finish first snapshot

+++ src/main/java/com/family/data/entity/Sku.java
@@ -37,6 +37,6 @@ public class Sku {
 	private Date createTime;
 	private Date updateTime;
 	private BigDecimal price;
-	private BigDecimal star;
+	private Integer star;
 
 }
